update translation tree
