+++ src/app/features/spaces/translations/shared/models/translation.model.ts
@@ -1,0 +1,5 @@
+export interface TranslationNode {
+  name: string;
+  key: string;
+  children?: TranslationNode[];
+}
